local/requirements.txt
aenum==3.1.15
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
dbf==0.99.9
et_xmlfile==2.0.0
fonttools==4.56.0
geopandas==1.0.1
idna==3.10
ifcopenshell==0.8.1.post1
isodate==0.7.2
kiwisolver==1.4.8
lark==1.2.2
lxml==5.3.1
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.2
openpyxl==3.1.5
osmnx==2.0.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.0
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.1
shapely==2.0.7
simpledbf==0.2.6
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
